Implement routing for EventDetail
